output summary count to ES when the job completes
