--- ProxyStatsGathering/src/main/scala/models/FinalCount.scala
@@ -1,3 +1,0 @@
-package models
-
-case class FinalCount (proxiedCount:Int, partialCount:Int, unProxiedCount:Int, notNeededCount:Int, dotFile:Int, glacier:Int)
