+++ common/src/main/scala/com/theguardian/multimedia/archivehunter/common/cmn_models/ProblemItemCount.scala
@@ -1,0 +1,5 @@
+package com.theguardian.multimedia.archivehunter.common.cmn_models
+
+import java.time.ZonedDateTime
+
+case class ProblemItemCount(scanStart:ZonedDateTime, scanFinish:Option[ZonedDateTime], proxiedCount:Int, partialCount:Int, unProxiedCount:Int, notNeededCount:Int, dotFile:Int, glacier:Int)
